+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-10-01
+## [Unreleased] - 2024-04-06
 
 ### Added
+* SortingNetwork class, with the following:
+  * compareExchange methods for arrays of doubles and ints
 
 ### Changed